requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
dependencies==1.0.1
dj-static==0.0.6
Django==3.0.6
djangorestframework==3.11.0
idna==2.9
mysqlclient==1.4.6
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
static3==0.7.0
urllib3==1.25.9
